refine export, fixtypo, add UT
