+++ auto_round/export/export_to_autoround/qlinear_fp.py
@@ -38,8 +38,7 @@ import transformers
 from auto_round.data_type.mxfp import FP32_EXPONENT_BIAS, FP32_MIN_NORMAL
 from auto_round.data_type.nvfp import cast_to_fp4, get_reciprocal
 from auto_round.data_type.utils import reshape_pad_tensor_by_group_size, revert_tensor_by_pad
-from auto_round.utils import is_mx_fp, is_nv_fp, is_standard_fp
-
+from auto_round.utils import is_nv_fp, is_mx_fp, is_standard_fp
 # from auto_round.utils import get_weight_compress_dtype
 logger = getLogger(__name__)
 E8M0_EXPONENT_BIAS = 127
