+++ test/test_cpu/requirements.txt
@@ -3,3 +3,4 @@ modelscope
 gguf
 mistral_common
 torchvision
+compressed-tensors
