+++ test/test_cpu/test_export.py
@@ -253,5 +253,170 @@ class TestAutoRound(unittest.TestCase):
         shutil.rmtree(quantized_model_path, ignore_errors=True)
 
 
+    def test_mxfp4_llmcompressor_format(self):
+        model_name = "facebook/opt-125m"
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
+        from transformers import AutoConfig
+        bits = 4
+        data_type = "mx_fp4e2m1"
+        group_size = 32
+        sym = True
+        autoround = AutoRound(
+            model,
+            self.tokenizer,
+            bits=bits,
+            group_size=group_size,
+            sym=sym,
+            iters=2,
+            seqlen=2,
+            data_type=data_type,
+            dataset=self.llm_dataloader,
+        )
+        quantized_model_path = self.save_dir
+        autoround.quantize()
+        compressed_model = autoround.save_quantized(
+            output_dir=quantized_model_path, inplace=True, format="llmcompressor"
+        )
+        tmp_layer = compressed_model.model.decoder.layers[3].self_attn.q_proj
+        assert (
+            hasattr(tmp_layer, "weight_scale")
+            and hasattr(tmp_layer, "weight_packed")
+            and tmp_layer.weight_scale.dtype is torch.uint8
+            and tmp_layer.weight_scale.shape[0] == 768
+        ), f"Illegal MXFP4 packing name or data_type or shape"
+        quantization_config = AutoConfig.from_pretrained(
+            quantized_model_path, trust_remote_code=True
+        ).quantization_config
+        assert (
+            quantization_config["format"] == "float-quantized"
+            and quantization_config["config_groups"]["group_0"]["weights"]["is_mx"] is True
+            and quantization_config["config_groups"]["group_0"]["weights"]["num_bits"] == 4
+        ), f"Invalid MXFP4 quantization configuration: {quantization_config}"
+        shutil.rmtree("./saved", ignore_errors=True)
+
+    def test_mxfp8_llmcompressor_format(self):
+        model_name = "facebook/opt-125m"
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
+        from transformers import AutoConfig
+        bits = 8
+        data_type = "mx_fp8e4m3_rceil"
+        group_size = 32
+        sym = True
+        autoround = AutoRound(
+            model,
+            self.tokenizer,
+            bits=bits,
+            group_size=group_size,
+            sym=sym,
+            iters=2,
+            seqlen=2,
+            data_type=data_type,
+            dataset=self.llm_dataloader,
+        )
+        quantized_model_path = self.save_dir
+        compressed_model,_ = autoround.quantize_and_save(
+            output_dir=quantized_model_path, format="llmcompressor"
+        )
+        tmp_layer = compressed_model.model.decoder.layers[3].self_attn.q_proj
+        assert (
+            hasattr(tmp_layer, "weight_scale")
+            and hasattr(tmp_layer, "weight")
+            and tmp_layer.weight.dtype is torch.float8_e4m3fn
+            and tmp_layer.weight_scale.dtype is torch.uint8
+            and tmp_layer.weight_scale.shape[0] == 768
+        ), f"Illegal MXFP8 packing name or data_type or shape"
+        quantization_config = AutoConfig.from_pretrained(
+            quantized_model_path, trust_remote_code=True
+        ).quantization_config
+        assert (
+            quantization_config["format"] == "float-quantized"
+            and quantization_config["config_groups"]["group_0"]["weights"]["is_mx"] is True
+            and quantization_config["config_groups"]["group_0"]["weights"]["num_bits"] == 8
+        ), f"Invalid MXFP8 quantization configuration: {quantization_config}"
+        shutil.rmtree("./saved", ignore_errors=True)
+
+    def test_nvfp4_llmcompressor_format(self):
+        model_name = "facebook/opt-125m"
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
+        from transformers import AutoConfig
+
+        bits = 4
+        act_bits = 4
+        data_type = "nv_fp"
+        act_data_type = "nv_fp4_with_static_gs"
+        group_size = 16
+        sym = True
+        autoround = AutoRound(
+            model,
+            self.tokenizer,
+            bits=bits,
+            act_bits=act_bits,
+            data_type=data_type,
+            act_data_type=act_data_type,
+            group_size=group_size,
+            sym=sym,
+            iters=2,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+        )
+        quantized_model_path = self.save_dir
+        compressed_model, _ = autoround.quantize_and_save(output_dir=quantized_model_path, format="llmcompressor")
+        tmp_layer = compressed_model.model.decoder.layers[3].self_attn.q_proj
+        assert (
+            hasattr(tmp_layer, "weight_scale")
+            and hasattr(tmp_layer, "weight_global_scale")
+            and hasattr(tmp_layer, "input_global_scale")
+            and tmp_layer.weight_packed.dtype is torch.uint8
+            and tmp_layer.weight_scale.dtype is torch.float8_e4m3fn
+            and tmp_layer.weight_scale.shape[0] == 768
+        ), "Illegal NVFP4 packing name or data_type or shape"
+        quantization_config = AutoConfig.from_pretrained(
+            quantized_model_path, trust_remote_code=True
+        ).quantization_config
+        assert (
+            quantization_config["format"] == "nvfp4-pack-quantized"
+            and quantization_config["config_groups"]["group_0"]["input_activations"]["num_bits"] == 4
+        ), f"Invalid NVFP4 quantization configuration: {quantization_config}"
+        shutil.rmtree("./saved", ignore_errors=True)
+
+    def test_nvfp4_autoround_format(self):
+        model_name = "facebook/opt-125m"
+        model = AutoModelForCausalLM.from_pretrained(model_name, torch_dtype="auto", trust_remote_code=True)
+        from transformers import AutoConfig
+
+        bits = 4
+        act_bits = 4
+        data_type = "nv_fp"
+        act_data_type = "nv_fp4_with_static_gs"
+        group_size = 16
+        sym = True
+        autoround = AutoRound(
+            model,
+            self.tokenizer,
+            bits=bits,
+            act_bits=act_bits,
+            data_type=data_type,
+            act_data_type=act_data_type,
+            group_size=group_size,
+            sym=sym,
+            iters=2,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+        )
+        quantized_model_path = self.save_dir
+        compressed_model, _ = autoround.quantize_and_save(output_dir=quantized_model_path, format="auto_round")
+        tmp_layer = compressed_model.model.decoder.layers[3].self_attn.q_proj
+        assert (
+            hasattr(tmp_layer, "weight_scale")
+            and hasattr(tmp_layer, "weight_global_scale")
+            and hasattr(tmp_layer, "input_global_scale")
+            and tmp_layer.weight_packed.dtype is torch.uint8
+            and tmp_layer.weight_scale.dtype is torch.float8_e4m3fn
+            and tmp_layer.weight_scale.shape[0] == 768
+        ), "Illegal NVFP4 packing name or data_type or shape"
+        shutil.rmtree("./saved", ignore_errors=True)
+
+
 if __name__ == "__main__":
     unittest.main()
+
