+++ test/test_cuda/requirements.txt
@@ -18,4 +18,6 @@ torchvision
 tqdm
 transformers
 vllm>=0.8.5.post1
+compressed-tensors
+
