+++ test/test_cuda/test_export.py
@@ -6,7 +6,7 @@ import unittest
 sys.path.insert(0, "../..")
 import torch
 import transformers
-from transformers import AutoModelForCausalLM, AutoTokenizer
+from transformers import AutoConfig, AutoModelForCausalLM, AutoTokenizer
 
 from auto_round import AutoRound
 from auto_round.testing_utils import require_awq, require_optimum
@@ -25,7 +25,7 @@ class TestAutoRound(unittest.TestCase):
     @classmethod
     def setUpClass(self):
         self.model_name = "facebook/opt-125m"
-
+        self.save_dir = "./saved"
         self.llm_dataloader = LLMDataLoader()
 
     @classmethod
@@ -320,6 +320,72 @@ class TestAutoRound(unittest.TestCase):
         print(tokenizer.decode(model.generate(**inputs, max_new_tokens=50)[0]))
         shutil.rmtree("./saved", ignore_errors=True)
 
+    def test_nvfp4_llmcompressor_format(self):
+        model = AutoModelForCausalLM.from_pretrained(self.model_name, torch_dtype="auto", trust_remote_code=True)
+        tokenizer = AutoTokenizer.from_pretrained(self.model_name, trust_remote_code=True)
+
+        bits = 4
+        act_bits = 4
+        data_type = "nv_fp"
+        act_data_type = "nv_fp4_with_static_gs"
+        group_size = 16
+        sym = True
+        autoround = AutoRound(
+            model,
+            tokenizer,
+            bits=bits,
+            act_bits=act_bits,
+            data_type=data_type,
+            act_data_type=act_data_type,
+            group_size=group_size,
+            sym=sym,
+            iters=2,
+            seqlen=2,
+            dataset=self.llm_dataloader,
+        )
+        quantized_model_path = self.save_dir
+        compressed_model, _ = autoround.quantize_and_save(output_dir=quantized_model_path, format="llmcompressor")
+        tmp_layer = compressed_model.model.decoder.layers[3].self_attn.q_proj
+        assert (
+            hasattr(tmp_layer, "weight_scale")
+            and hasattr(tmp_layer, "weight_global_scale")
+            and hasattr(tmp_layer, "input_global_scale")
+            and tmp_layer.weight_packed.dtype is torch.uint8
+            and tmp_layer.weight_scale.dtype is torch.float8_e4m3fn
+            and tmp_layer.weight_scale.shape[0] == 768
+        ), "Illegal NVFP4 packing name or data_type or shape"
+        quantization_config = AutoConfig.from_pretrained(
+            quantized_model_path, trust_remote_code=True
+        ).quantization_config
+        assert (
+            quantization_config["format"] == "nvfp4-pack-quantized"
+            and quantization_config["config_groups"]["group_0"]["input_activations"]["num_bits"] == 4
+        ), f"Invalid NVFP4 quantization configuration: {quantization_config}"
+        shutil.rmtree("./saved", ignore_errors=True)
+        # from vllm import LLM, SamplingParams
+        # prompts = [
+        #     "The capital of France is",
+        #     "The future of AI is",
+        # ]
+        ## Create a sampling params object.
+        # sampling_params = SamplingParams(temperature=0.8, top_p=0.95, max_tokens=20)
+        # QUANTIZATION = "compressed-tensors"
+        # llm = LLM(model=quantized_model_path,
+        #         #   quantization=QUANTIZATION,
+        #           trust_remote_code=True,
+        #           tensor_parallel_size=1,
+        #           enforce_eager=True,
+        #           gpu_memory_utilization=0.7,
+        # )
+        # outputs = llm.generate(prompts, sampling_params)
+        # # Print the outputs.
+        # for output in outputs:
+        #     prompt = output.prompt
+        #     generated_text = output.outputs[0].text
+        #     if "France" in prompt:
+        #         assert "Paris" in generated_text
+
 
 if __name__ == "__main__":
     unittest.main()
+
